Factor IncidentReportCard out of IncidentReqportTracker
